+++ src/main/java/com/airbus_cyber_security/graylog/wizard/alert/rest/models/requests/AlertRuleRequest.java
@@ -40,9 +40,8 @@ public abstract class AlertRuleRequest {
     @NotNull
     public abstract String getSeverity();
 
-    // TODO should be @NotNull (encode the absence of description with an empty string)
     @JsonProperty("description")
-    @Nullable
+    @NotNull
     public abstract String getDescription();
 
     // TODO should be @NotNull