+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [4.5.0](https://github.com/airbus-cyber/graylog-plugin-alert-wizard/compare/4.4.1...4.5.0)
 ### Features
+* the rule description is now mapped to the event definition description ([issue #102](https://github.com/airbus-cyber/graylog-plugin-alert-wizard/issues/102))
 * POST plugins/com.airbus_cyber_security.graylog.wizard/alerts now returns data about the newly created alert rule
 * resource plugins/com.airbus_cyber_security.graylog.wizard/alerts returns the second event definition identifier (if any) as field `second_event_definition`
 ### Bug Fixes