+++ CHANGELOG.md
@@ -8,12 +8,12 @@ All notable changes to this project will be documented in this file.
 * POST plugins/com.airbus_cyber_security.graylog.wizard/alerts now returns data about the newly created alert rule
 * Resource plugins/com.airbus_cyber_security.graylog.wizard/alerts returns the second event definition identifier (if any) as field `second_event_definition`
 * Disabling a wizard rule now disables the associated event definitions ([issue #58](https://github.com/airbus-cyber/graylog-plugin-alert-wizard/issues/58))
-* Update graylog-plugin-logging-alert dependency to [4.4](https://github.com/airbus-cyber/graylog-plugin-logging-alert/blob/4.4.1/CHANGELOG.md)
 ### Bug Fixes
 * After deleting the stream associated with a rule, the list of rules would not load anymore ([issue #105](https://github.com/airbus-cyber/graylog-plugin-alert-wizard/issues/105))
 * The backlog size value set in the configuration is now taken into account in the event definition ([issue #40](https://github.com/airbus-cyber/graylog-plugin-alert-wizard/issues/40))
 * Several display bugs in dark mode fixed ([issue #59](https://github.com/airbus-cyber/graylog-plugin-alert-wizard/issues/59))
 * French translation of statistical condition 'sum' was incorrect ([issue #112](https://github.com/airbus-cyber/graylog-plugin-alert-wizard/issues/112))
+* Update graylog-plugin-logging-alert dependency to [4.4.1](https://github.com/airbus-cyber/graylog-plugin-logging-alert/blob/4.4.1/CHANGELOG.md)
 ### Changes
 * POST plugins/com.airbus_cyber_security.graylog.wizard/alerts, field `description` should be not null (use an empty string, if there is no description)